+++ lib/rubocop/cop/open_project_cops.rb
@@ -2,5 +2,6 @@
 
 require_relative "open_project/add_preview_for_view_component"
 require_relative "open_project/no_not_implemented_error"
+require_relative "open_project/no_params_in_work_package_where_id"
 require_relative "open_project/use_service_result_factory_methods"
 require_relative "open_project/no_sleep_in_feature_specs"